requirements.txt
flask==2.3.3
gunicorn==21.2.0
google-cloud-bigquery==3.11.4
requests==2.31.0
google-cloud-logging==3.8.0
